PBI-3503 Group Message list by date (#309)
* PBI-3503 Group message list by date
[PBI-3503]

* Move grouping to viewmodel.
These features shouldn't be stored anywhere.
Just evaluate and show on UI as needed.

* Remove null setting

* ktlint

* Write up unit tests

* Convert to millis

* Fix tests

* Have groupings show first time as well.

* Group new messages

* Update new grouped messages

* Use seconds off the bat. Sever returns seconds instead of milliseconds. Comments.

* ktlint

+++ apptentive-core/src/main/java/apptentive/com/android/network/HttpNetwork.kt
@@ -82,8 +82,8 @@ class DefaultHttpNetwork(
      */
     private fun openConnection(request: HttpRequest<*>): HttpURLConnection {
         val connection = createConnection(request.url)
-        connection.connectTimeout = toMilliseconds(connectTimeout)
-        connection.readTimeout = toMilliseconds(readTimeout)
+        connection.connectTimeout = toMilliseconds(connectTimeout).toInt()
+        connection.readTimeout = toMilliseconds(readTimeout).toInt()
         connection.useCaches = false
         connection.doInput = true
         return connection
